src/main/webapp/app/entities/field-test-entity/update/field-test-entity-update.component.spec.ts
jest.mock('@angular/router');

import { ComponentFixture, TestBed, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { of } from 'rxjs';

import { FieldTestEntityService } from '../service/field-test-entity.service';
import { FieldTestEntity } from '../field-test-entity.model';

import { FieldTestEntityUpdateComponent } from './field-test-entity-update.component';

describe('Component Tests', () => {
  describe('FieldTestEntity Management Update Component', () => {
    let comp: FieldTestEntityUpdateComponent;
    let fixture: ComponentFixture<FieldTestEntityUpdateComponent>;
    let service: FieldTestEntityService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [FieldTestEntityUpdateComponent],
        providers: [FormBuilder, ActivatedRoute],
      })
        .overrideTemplate(FieldTestEntityUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(FieldTestEntityUpdateComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(FieldTestEntityService);
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', fakeAsync(() => {
        // GIVEN
        const entity = new FieldTestEntity('123');
        spyOn(service, 'update').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.update).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));

      it('Should call create service on save for new entity', fakeAsync(() => {
        // GIVEN
        const entity = new FieldTestEntity();
        spyOn(service, 'create').and.returnValue(of(new HttpResponse({ body: entity })));
        comp.updateForm(entity);
        // WHEN
        comp.save();
        tick(); // simulate async

        // THEN
        expect(service.create).toHaveBeenCalledWith(entity);
        expect(comp.isSaving).toEqual(false);
      }));
    });
  });
});
